% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AnM}
\alias{AnM}
\title{Compute Lipschitz bound upper bound on the largest eigenvalue of the Laplacian}
\usage{
AnM(w, p)
}
\arguments{
\item{w}{vector of weights}

\item{p}{number of points to cluster}
}
\description{
\code{AnM} computes the node degrees from the edge information.
}
\examples{
data(mammals)
X <- as.matrix(mammals[,-1])
X <- t(scale(X,center=TRUE,scale=FALSE))
p <- ncol(X)

## Pick some weights and a sequence of regularization parameters.
k <- 5
phi <- 0.5
w <- kernel_weights(X,phi)
w <- knn_weights(w,k,p)
AnM(w,p)
}

