\name{kernel_weights}
\alias{kernel_weights}
\title{Compute Gaussian Kernel Weights}
\usage{
  kernel_weights(X, mu)
}
\arguments{
  \item{X}{q-by-p data matrix}

  \item{mu}{nonnegative parameter that controls the scale
  of kernel weights}
}
\description{
  \code{kernel_weights} computes Gaussian kernel weights
}

