\docType{data}
\name{mammals}
\alias{mammals}
\title{Tally of types of teeth in some mammals.}
\format{A data frame with 27 rows and 8 variables}
\description{
  A dataset containing how many of eight kinds of teeth
  various mammals have. This data set is a subset of the
  mammals dentition data in the homals package.
}
\references{
  Jan de Leeuw, Patrick Mair (2009). Gifi Methods for
  Optimal Scaling in R: The Package homals. Journal of
  Statistical Software, August 2009, Volume 31, Issue 4.
  \url{http://www.jstatsoft.org/v31/i04/paper}.
}
\keyword{datasets}

