\docType{package}
\name{cvxclustr}
\alias{cvxclustr}
\alias{cvxclustr-package}
\title{Convex Clustering via Splitting Methods}
\description{
  AMA and ADMM splitting methods for convex clustering.
}
\references{
  Eric C. Chi, Kenneth Lange (2013). Splitting Methods for
  convex clustering. arXiv: 1304.0499 [stat.ML].
  \url{http://arxiv.org/abs/1304.0499}.
}

