% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_clusters}
\alias{find_clusters}
\title{Find clusters}
\usage{
find_clusters(A)
}
\arguments{
\item{A}{adjacency matrix}
}
\description{
\code{find_clusters} uses breadth-first search to identify the connected components of the corresponding
adjacency graph of the centroid differences vectors.
}

